add SaveInvoiceAsync method

+++ src/TaskrowSharp/Models/InvoiceFeeeDetailResponse.cs
@@ -3,6 +3,7 @@
     public class InvoiceFeeeDetailResponse
     {
         public InvoiceFee InvoiceFee { get; set; }
-        public double MonthRevenueValue { get; set; }
+
+        public decimal MonthRevenueValue { get; set; }
     }
 }
